adapt new api

+++ src/mindytool/data/MapData.java
@@ -14,6 +14,6 @@ public class MapData {
     Long dislikes;
     Boolean isCurated;
     Boolean isVerified;
-    Integer downloadCount;
+    Long downloadCount;
     Boolean isPrivate;
 }
